add verifyToken middleware
